requirements.txt
Django>=5.2,<6.1
djangorestframework>=3.14.0
requests>=2.31.0
geopy>=2.4.1
polyline>=2.0.1